requirements.txt
Django==2.0.4
requests==2.18.4
python-dateutil==2.7.2
